stocks/requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astunparse==1.6.3
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-crontab==0.7.1
django-rest-auth==0.9.5
django-rest-registration==0.5.6
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
joblib==0.15.1
Keras==2.4.2
Keras-Preprocessing==1.1.2
lxml==4.5.1
Markdown==3.2.2
mysqlclient==1.4.6
nltk==3.5
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.5
pandas-datareader==0.8.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.47.0
typing==3.7.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
